+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `shopify/sync` commands now support a `--throttle` option.
 - Fixed a bug where syncing Shopify variants would be limited to 50. ([#115](https://github.com/craftcms/shopify/issues/115))
+- Fixed a bug where variant column data could be larger than a MySQL TEXT column.
 - Added `craft\shopify\console\controllers\SyncController::$throttle`.
 - Added `craft\shopify\services\Products::$throttle`.
 - Added `craft\shopify\services\Products::$sleepSeconds`.